requirements.txt
asgiref==3.11.1
autopep8==2.3.2
cffi==2.0.0
cryptography==46.0.6
dj-database-url==2.3.0
Django==6.0
django-extensions==4.1
djangorestframework==3.14.0
flake8==7.3.0
gunicorn==25.3.0
MarkupSafe==3.0.3
mccabe==0.7.0
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pycparser==3.0
pyflakes==3.4.0
pyOpenSSL==26.0.0
python-decouple==3.8
python-dotenv==1.2.2
sqlparse==0.5.5
supabase==2.15.0
typing_extensions==4.15.0
Werkzeug==3.1.8
whitenoise==6.9.0
django-ratelimit==4.1.0
